Helper scripts for seuronbot
Because chunkflow is extremely flexible, it is difficult to create
a seuronbot interface to do everything possible. The current
approach is to only have a minimal subset of parameters actually go
through the bot, like input output location, bounding box etc, the
rest are handled by the helper scripts, one can either hard code
the parameters in the scripts, or read the parameters from the json
file. Seuronbot will pass all the content in the input json file to
the worker, so it is possible to define one's own seuronbot
interface as long as one pass the matching chunkflow image